Se detecta cuando el usuario toca sobre un boton. (back, A, B)
