--- core/src/mx/tec/astral/flytomars/AlienAgil.java
@@ -1,9 +1,0 @@
-package mx.tec.astral.flytomars;
-
-import com.badlogic.gdx.graphics.Texture;
-
-public class AlienAgil extends Objeto{
-    public AlienAgil(Texture textura, float x, float y){
-        super(textura, x, y);
-    }
-}
