--- core/src/mx/tec/astral/flytomars/AlienLetal.java
@@ -1,10 +1,0 @@
-package mx.tec.astral.flytomars;
-
-import com.badlogic.gdx.graphics.Texture;
-
-public class AlienLetal extends Objeto{
-    public AlienLetal(Texture textura, float x, float y){
-        super(textura,x,y);
-
-    }
-}
