--- core/src/mx/tec/astral/flytomars/AlienTanque.java
@@ -1,9 +1,0 @@
-package mx.tec.astral.flytomars;
-
-import com.badlogic.gdx.graphics.Texture;
-
-public class AlienTanque extends Objeto{
-    public AlienTanque(Texture textura, float x, float y){
-        super(textura, x, y);
-    }
-}
