--- core/src/mx/tec/astral/flytomars/PantallaNvl2.java
@@ -1,42 +1,0 @@
-package mx.tec.astral.flytomars;
-
-import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.graphics.g2d.BitmapFont;
-
-public class PantallaNvl2 extends Pantalla {
-    private Juego juego;
-    BitmapFont font = new BitmapFont(); //or use alex answer to use custom font
-
-    public PantallaNvl2(Juego juego) {
-        this.juego = juego;
-    }
-
-    @Override
-    public void show() {
-
-    }
-
-    @Override
-    public void render(float delta) {
-        borrarPantalla(0,0,0);
-        batch.begin();
-        font.getData().setScale(3,3);
-        font.draw(batch, "Pantalla Nivel 2:", 25, ALTO - 25 );
-        batch.end();
-    }
-
-    @Override
-    public void pause() {
-
-    }
-
-    @Override
-    public void resume() {
-
-    }
-
-    @Override
-    public void dispose() {
-
-    }
-}
