+++ core/src/mx/tec/astral/flytomars/Pantallas/PantallaJuego.java
@@ -8,7 +8,6 @@ import com.badlogic.gdx.scenes.scene2d.ui.Button;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.badlogic.gdx.utils.Align;
-
 import mx.tec.astral.flytomars.Juego;
 
 //import sun.security.util.AnchorCertificates;
