+++ core/src/mx/tec/astral/flytomars/Pantallas/PantallaNvl1.java
@@ -16,6 +16,7 @@ import com.badlogic.gdx.utils.Array;
 import mx.tec.astral.flytomars.Enemigos.AlienAgil;
 import mx.tec.astral.flytomars.Enemigos.AlienLetal;
 import mx.tec.astral.flytomars.Enemigos.AlienTanque;
+import mx.tec.astral.flytomars.EstadoHeroe;
 import mx.tec.astral.flytomars.Hero;
 import mx.tec.astral.flytomars.Juego;
 import mx.tec.astral.flytomars.Vida;
@@ -25,8 +26,8 @@ Pantalla que almacena todos los objetos del nivel 1
 Autores: Israel, Misael y Alejandro
  */
 public class PantallaNvl1 extends Pantalla {
-    // Font of score.
-    BitmapFont font = new BitmapFont(); //or use alex answer to use custom font
+// Font of score.
+BitmapFont font = new BitmapFont(); //or use alex answer to use custom font
 
     //Velocidad del Hero
     private static  final float DELTA_X_HERO = 10;
@@ -35,11 +36,11 @@ public class PantallaNvl1 extends Pantalla {
     Texture texturaFondo;
     private Stage escenaMenuNiveles;
 
-    //Personaje (Hero)
-    private Hero heroR;
-    private  Hero heroL;
+    //  Personaje (Hero)
+    private  Hero hero;
+    private EstadoHeroe prevState = EstadoHeroe.DERECHA;
 
-    //Enemigos
+    //  Enemigos
     //Alien Agil
     private AlienAgil aAgil;
     //Alien Letal
@@ -48,17 +49,18 @@ public class PantallaNvl1 extends Pantalla {
     private AlienTanque aTanque;
 
 
-    //Objetos vida
+    //  Objetos vida
     private Array<Vida> arrVidas;
+
+    //  Buttons
+    private Texture texturaBack;
+    private Texture texturaA;
+    private Texture texturaB;
+
 
     //Indican si el Hero se mueve en cierta dirección
     private boolean moviendoIzquierda = false;
     private boolean moviendoDerecha = false;
-    private boolean posDerecha = true;
-    private boolean posIzquierda = false;
-
-    //Checa si el juego se pauso
-    boolean paused;
 
 
     public PantallaNvl1(Juego juego) {
@@ -67,15 +69,44 @@ public class PantallaNvl1 extends Pantalla {
 
     @Override
     public void show() {
-        crearMenu();
-        heroR = crearHero(heroR, false);
-        heroL = crearHero(heroL, true);
+
+        crearFondo();
+
+        crearHero();
         crearAlienAgil();
         crearAlienLetal();
         crearAlienTanque();
         crearVidas();
+
+        crearBotonBack();
+        crearBotonA();
+        crearBotonB();
+
         //Ahora la misma pantalla RECIBE Y PROCESA los eventos
         Gdx.input.setInputProcessor(new ProcesadorEntrada());
+    }
+
+    private void crearFondo() {
+        texturaFondo = new Texture("nivel1/lvlEarth.png");
+    }
+
+    private void crearBotonA() {
+        texturaA = new Texture("buttons/btn_A.png");
+    }
+
+    private void crearBotonB() {
+        texturaB = new Texture("buttons/btn_B.png");
+    }
+
+    private void crearBotonBack() {
+        texturaBack = new Texture("Menu/btn_back.png");
+    }
+
+    private void crearHero() {
+        Texture texturaIzquierda = new Texture("nivel1/character1_left.png");
+        Texture texturaDerecha = new Texture("nivel1/character1.png");
+
+        hero = new Hero(texturaDerecha, texturaIzquierda, 0, 150);
     }
 
     private void crearAlienTanque() {
@@ -98,90 +129,19 @@ public class PantallaNvl1 extends Pantalla {
         Texture texturaVida = new Texture("items/heart.png");
 
         arrVidas = new Array<>(3);
-            for (int i = 0; i < 3; i++){
-                Vida vida = new Vida(texturaVida,  ANCHO-(i*60+65),ALTO-60);
-                arrVidas.add(vida); //Lo guarda en el arrelo
-            }
+        for (int i = 0; i < 3; i++){
+            Vida vida = new Vida(texturaVida,  ANCHO-(i*60+65),ALTO-60);
+            arrVidas.add(vida); //Lo guarda en el arrelo
+        }
     }
 
-    //Crea la imagen de nuestro protagonista
-    private Hero crearHero(Hero myHero, boolean isLeft) {
-        Texture texturaHero;
-        if(!isLeft)
-             texturaHero = new Texture("nivel1/character1.png");
-        else
-            texturaHero = new Texture("nivel1/character1_left.png");
-        myHero = new Hero(texturaHero,0,150);
-        return myHero;
-    }
-
-    private void crearMenu() {
-        texturaFondo = new Texture("nivel1/lvlEarth.png");
-
-
-        // MENU, necesitamos una escena
-        //Escena
-        escenaMenuNiveles = new Stage(vista);
-
-        // Actores->Boton
-        Button btnA = crearBoton("buttons/btn_A.png", "buttons/btn_A_press.png");
-        Button btnB = crearBoton("buttons/btn_B.png", "buttons/btn_B_press.png");
-        Button btnBack = crearBoton("Menu/btn_back.png", "Menu/btn_back_press.png");
-
-
-        //Se les agrega una posicion en pantalla
-        btnA.setPosition(ANCHO-btnA.getWidth()*2, 100, Align.center);
-        btnB.setPosition(ANCHO-btnB.getWidth(), 100, Align.center);
-        btnBack.setPosition(ANCHO/4 - btnBack.getMinWidth(), 100, Align.center);
-
-        // Agrega los botones a escena
-        escenaMenuNiveles.addActor(btnA);
-        escenaMenuNiveles.addActor(btnB);
-        escenaMenuNiveles.addActor(btnBack);
-
-        //Aqui se agrega la accion a ejecutar cuando se presiona
-        btnA.addListener(new ClickListener(){
-            @Override
-            public void clicked(InputEvent event, float x, float y) {
-                //juego.setScreen(new PantallaNvl1(juego));
-            }
-        });
-
-        btnB.addListener(new ClickListener(){
-            @Override
-            public void clicked(InputEvent event, float x, float y) {
-                //juego.setScreen(new PantallaNvl2(juego));
-            }
-        });
-
-        btnBack.addListener(new ClickListener(){
-            @Override
-            public void clicked(InputEvent event, float x, float y) {
-                juego.setScreen(new PantallaMenu(juego));
-            }
-        });
-
-
-        // La ESCENA se encarga de ATENDER LOS EVENTOS DE ENTRADA
-        Gdx.input.setInputProcessor(escenaMenuNiveles);
-
-    }
-
-    private Button crearBoton(String archivo, String clickeado) {
-        Texture texturaBoton = new Texture(archivo);
-        TextureRegionDrawable trdBtn = new TextureRegionDrawable(texturaBoton);
-        // Clickeado
-        Texture texturaClickeada = new Texture(clickeado);
-        TextureRegionDrawable trdBtnClick = new TextureRegionDrawable(texturaClickeada);
-        return new Button(trdBtn, trdBtnClick);
-    }
     @Override
     public void render(float delta) {
-        actualizar(heroL);
-        actualizar(heroR);
+//        actualizar(heroL);
+//        actualizar(heroR);
+        actualizar(hero);
         borrarPantalla(0,0,0); //Borrar con color negro}
         batch.setProjectionMatrix(camara.combined);
-        camara.update();
 
         batch.begin();
 
@@ -199,10 +159,7 @@ public class PantallaNvl1 extends Pantalla {
         }
 
         //Hero
-        if(moviendoDerecha || posDerecha)
-            heroR.render(batch);
-        else if(moviendoIzquierda || posIzquierda)
-            heroL.render(batch);
+        hero.render(batch);
 
         //Enemigos
         //Alien Agil
@@ -212,25 +169,38 @@ public class PantallaNvl1 extends Pantalla {
         //Alien Tanque
         aTanque.render(batch);
 
+        // Draw A
+        batch.draw(texturaA, ANCHO-texturaA.getWidth()*2, texturaA.getHeight()/2f);
+
+        // Draw B
+        batch.draw(texturaB, ANCHO-texturaB.getWidth(), texturaB.getHeight()/2f);
+
+        // Draw Back
+        batch.draw(texturaBack, texturaBack.getWidth()/2f, texturaBack.getHeight());
+
         batch.end();
 
-
-
-        escenaMenuNiveles.draw();
     }
 
     private void actualizar(Hero hero) {
-        if(moviendoDerecha && hero.getX() <= (ANCHO- hero.getWidth())){
+        if(prevState == EstadoHeroe.DERECHA && hero.getEstado() == EstadoHeroe.IZQUIERDA)
+            hero.cambiarEstado();
+        else if(prevState == EstadoHeroe.IZQUIERDA && hero.getEstado() == EstadoHeroe.DERECHA)
+            hero.cambiarEstado();
+
+        if(moviendoDerecha && hero.getX() <= (ANCHO- hero.getWidth())) {
             hero.mover(DELTA_X_HERO);
-        }if(moviendoIzquierda && hero.getX() > 0)
-        {
+            prevState = EstadoHeroe.DERECHA;
+        }if(moviendoIzquierda && hero.getX() > 0) {
             hero.mover(-DELTA_X_HERO);
+            prevState = EstadoHeroe.IZQUIERDA;
         }
+
+
     }
 
     @Override
-    public void pause()
-    {
+    public void pause() {
 
     }
 
@@ -269,17 +239,29 @@ public class PantallaNvl1 extends Pantalla {
             //button = dice si presione el boton izq o derecho
             Vector3 v = new Vector3(screenX,screenY,0);
             camara.unproject(v); //Convierte de coordenadas FISICAS a LÓGICAS
-            if(v.x < ANCHO/2){
-                //Primera mitad de la pantalla
-                //nave.mover(-DELTA_X);
-                moviendoIzquierda = true;
-                posDerecha = false;
-                posIzquierda = true;
-            }else{
-                heroR.mover(DELTA_X_HERO);
-                moviendoDerecha = true;
-                posDerecha = true;
-                posIzquierda = false;
+
+            // Back button
+            if(v.x >= texturaBack.getWidth()/2f && v.x <= texturaBack.getWidth()*1.5f &&
+                    v.y >= texturaBack.getHeight() && v.y <= texturaBack.getHeight()*2)
+                juego.setScreen(new PantallaJuego(juego));
+
+                //  A button (Jump)
+            else if(v.x >= ANCHO-texturaA.getWidth()*2 && v.x <=ANCHO-texturaA.getWidth() &&
+                    v.y >= texturaA.getHeight()/2f && v.y <= texturaA.getHeight()*1.5f)
+                Gdx.app.log("A_button", "A pressed!");
+
+                //  B button (Shoot)
+            else if(v.x >= ANCHO-texturaB.getWidth() && v.x <= ANCHO &&
+                    v.y >= texturaB.getHeight()/2f && v.y <= texturaB.getHeight()*1.5f)
+                Gdx.app.log("B_button", "B pressed!");
+
+            else{
+                if(v.x < ANCHO/2){
+                    //Primera mitad de la pantalla
+                    moviendoIzquierda = true;
+                }else{
+                    moviendoDerecha = true;
+                }
             }
             return true; //Porque el juego ya proceso el evento
         }
